refactor: Rename and break components into smaller components
