+++ app/ai/components/suggested-pill-actions.tsx
@@ -5,7 +5,7 @@ import { PROMPTS } from "@/utils/prompts";
 import { createCurrentThemePrompt } from "@/utils/ai-prompt";
 import { AIPromptData } from "@/types/ai";
 import { Sparkles } from "lucide-react";
-import { AIPillActionButton } from "@/components/editor/ai/ai-pill-action-button";
+import { PillActionButton } from "@/components/editor/ai/pill-action-button";
 
 export function SuggestedPillActions({
   handleThemeGeneration,
@@ -22,13 +22,13 @@ export function SuggestedPillActions({
   return (
     <>
       {Object.entries(PROMPTS).map(([key, { label, prompt }]) => (
-        <AIPillActionButton
+        <PillActionButton
           key={key}
           onClick={() => handleSetPrompt(prompt)}
           disabled={aiIsGenerating}
         >
           <Sparkles /> {label}
-        </AIPillActionButton>
+        </PillActionButton>
       ))}
     </>
   );
